turn cuprated into a embeddable library
